[CORDA-1941]: Some TODOs to address.
